+++ datastore/couchDB/pouchDBServiceDatastoreCouchDB.go
@@ -340,6 +340,30 @@ func fetchRevDiff(body map[string]interface{}, resource *couchdb.Resource) (map[
 	return parseData(data)
 }
 
+func addDesignDocument(body map[string]interface{}, resource *couchdb.Resource) (map[string]interface{}, error) {
+	_, data, err := resource.PutJSON("", nil, body, nil)
+	if err != nil {
+		return nil, err
+	}
+
+	return parseData(data)
+}
+
+func fetchDesignDocumentResults(body map[string]interface{}, dbName string, indexName string) (map[string]interface{}, error) {
+	resource, err := couchdb.NewResource(dbName, nil)
+	if err != nil {
+		logger.Log.Debugf("Falied to create resource to DB %s: %s", dbName, err.Error())
+		return nil, err
+	}
+
+	_, data, err := resource.PostJSON(indexName, nil, body, nil)
+	if err != nil {
+		return nil, err
+	}
+
+	return parseData(data)
+}
+
 func performBulkUpdate(body map[string]interface{}, resource *couchdb.Resource) ([]map[string]interface{}, error) {
 	_, data, err := resource.PostJSON("_bulk_docs", nil, body, nil)
 	if err != nil {