+++ gathergrpc/gather.pb.gw.go
@@ -979,6 +979,37 @@ func request_TenantProvisioningService_GetAllMonitoredObjects_0(ctx context.Cont
 
 }
 
+func request_TenantProvisioningService_GetMonitoredObjectToDomainMap_0(ctx context.Context, marshaler runtime.Marshaler, client TenantProvisioningServiceClient, req *http.Request, pathParams map[string]string) (proto.Message, runtime.ServerMetadata, error) {
+	var protoReq MonitoredObjectCountByDomainRequest
+	var metadata runtime.ServerMetadata
+
+	if err := marshaler.NewDecoder(req.Body).Decode(&protoReq); err != nil {
+		return nil, metadata, status.Errorf(codes.InvalidArgument, "%v", err)
+	}
+
+	var (
+		val string
+		ok  bool
+		err error
+		_   = err
+	)
+
+	val, ok = pathParams["tenantId"]
+	if !ok {
+		return nil, metadata, status.Errorf(codes.InvalidArgument, "missing parameter %s", "tenantId")
+	}
+
+	protoReq.TenantId, err = runtime.String(val)
+
+	if err != nil {
+		return nil, metadata, status.Errorf(codes.InvalidArgument, "type mismatch, parameter: %s, error: %v", "tenantId", err)
+	}
+
+	msg, err := client.GetMonitoredObjectToDomainMap(ctx, &protoReq, grpc.Header(&metadata.HeaderMD), grpc.Trailer(&metadata.TrailerMD))
+	return msg, metadata, err
+
+}
+
 var (
 	filter_MetricsService_GetThresholdCrossing_0 = &utilities.DoubleArray{Encoding: map[string]int{}, Base: []int(nil), Check: []int(nil)}
 )
@@ -2093,6 +2124,35 @@ func RegisterTenantProvisioningServiceHandlerClient(ctx context.Context, mux *ru
 
 	})
 
+	mux.Handle("POST", pattern_TenantProvisioningService_GetMonitoredObjectToDomainMap_0, func(w http.ResponseWriter, req *http.Request, pathParams map[string]string) {
+		ctx, cancel := context.WithCancel(req.Context())
+		defer cancel()
+		if cn, ok := w.(http.CloseNotifier); ok {
+			go func(done <-chan struct{}, closed <-chan bool) {
+				select {
+				case <-done:
+				case <-closed:
+					cancel()
+				}
+			}(ctx.Done(), cn.CloseNotify())
+		}
+		inboundMarshaler, outboundMarshaler := runtime.MarshalerForRequest(mux, req)
+		rctx, err := runtime.AnnotateContext(ctx, mux, req)
+		if err != nil {
+			runtime.HTTPError(ctx, mux, outboundMarshaler, w, req, err)
+			return
+		}
+		resp, md, err := request_TenantProvisioningService_GetMonitoredObjectToDomainMap_0(rctx, inboundMarshaler, client, req, pathParams)
+		ctx = runtime.NewServerMetadataContext(ctx, md)
+		if err != nil {
+			runtime.HTTPError(ctx, mux, outboundMarshaler, w, req, err)
+			return
+		}
+
+		forward_TenantProvisioningService_GetMonitoredObjectToDomainMap_0(ctx, mux, outboundMarshaler, w, req, resp, mux.GetForwardResponseOptions()...)
+
+	})
+
 	return nil
 }
 
@@ -2142,6 +2202,8 @@ var (
 	pattern_TenantProvisioningService_GetMonitoredObject_0 = runtime.MustPattern(runtime.NewPattern(1, []int{2, 0, 2, 1, 2, 2, 1, 0, 4, 1, 5, 3, 2, 4, 1, 0, 4, 1, 5, 5}, []string{"api", "v1", "tenants", "tenantId", "monitored-objects", "monitoredObjectId"}, ""))
 
 	pattern_TenantProvisioningService_GetAllMonitoredObjects_0 = runtime.MustPattern(runtime.NewPattern(1, []int{2, 0, 2, 1, 2, 2, 1, 0, 4, 1, 5, 3, 2, 4}, []string{"api", "v1", "tenants", "value", "monitored-object-list"}, ""))
+
+	pattern_TenantProvisioningService_GetMonitoredObjectToDomainMap_0 = runtime.MustPattern(runtime.NewPattern(1, []int{2, 0, 2, 1, 2, 2, 1, 0, 4, 1, 5, 3, 2, 4}, []string{"api", "v1", "tenants", "tenantId", "monitored-object-domain-map"}, ""))
 )
 
 var (
@@ -2190,6 +2252,8 @@ var (
 	forward_TenantProvisioningService_GetMonitoredObject_0 = runtime.ForwardResponseMessage
 
 	forward_TenantProvisioningService_GetAllMonitoredObjects_0 = runtime.ForwardResponseMessage
+
+	forward_TenantProvisioningService_GetMonitoredObjectToDomainMap_0 = runtime.ForwardResponseMessage
 )
 
 // RegisterMetricsServiceHandlerFromEndpoint is same as RegisterMetricsServiceHandler but